+++ ssl_des/src/cc/des/des_cfb.c
@@ -6,7 +6,7 @@
 /*   By: hypark <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/28 23:01:33 by hypark            #+#    #+#             */
-/*   Updated: 2019/09/04 05:02:01 by hypark           ###   ########.fr       */
+/*   Updated: 2019/09/08 16:49:09 by hypark           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 static void				process_des_cfb(t_ssl *ssl, t_des *des)
 {
 	uint64_t			*m;
-	uint8_t				block_n;
+	uint64_t			block_n;
 	uint64_t			subkey[16];
 
 	des_generate_subkey(subkey, des->key);
